backend/requirements.txt
openai
python-dotenv
python-dotenv
Pillow
requests
